backend/pyproject.toml
[project]
name = "retreat-planner-backend"
version = "0.1.0"
description = "Multi-agent retreat planning backend with CrewAI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "crewai>=0.28.0",
    "crewai-tools>=0.12.0",
    "tavily-python>=0.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "stripe>=7.0.0",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
